+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-----
+### Added
+
+- Storage encryption at rest support via `encrypted`
 
 ## [2.5.1] - 2023-09-19